backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
alembic==1.15.2
pydantic==2.11.3
pydantic-settings==2.8.1
python-multipart==0.0.20
pyyaml==6.0.2
